refactor(project): 重构项目并支持 TypeScript - 将 .mts 文件重命名为 .ts - 更新导入路径以使用 TypeScript - 在 package.json 中添加 TypeScript 依赖- 更新 tsconfig.json 以包含新文件
