+++ src/frontend/src/main.ts
@@ -1,5 +1,5 @@
 import App from './App.vue'
-import router from './router.mjs'
+import router from './router.ts'
 import {createApp} from 'vue'
 
 createApp(App).use(router).mount('#app')
